Refactor DataCollector html dumper
