+++ src/DebugBar/DataCollector/DataCollector.php
@@ -10,69 +10,21 @@
 
 namespace DebugBar\DataCollector;
 
-use DebugBar\DataFormatter\DataFormatter;
-use DebugBar\DataFormatter\DataFormatterInterface;
-use DebugBar\DataFormatter\DebugBarVarDumper;
+use DebugBar\DataFormatter\HasDataFormatter;
 
 /**
  * Abstract class for data collectors
  */
 abstract class DataCollector implements DataCollectorInterface
 {
-    private static $defaultDataFormatter;
-    private static $defaultVarDumper;
+    use HasDataFormatter;
 
-    protected $dataFormater;
-    protected $varDumper;
+    public static $defaultDataFormatter;
+    public static $defaultVarDumper;
+
     protected $xdebugLinkTemplate = '';
     protected $xdebugShouldUseAjax = false;
     protected $xdebugReplacements = array();
-
-    /**
-     * Sets the default data formater instance used by all collectors subclassing this class
-     *
-     * @param DataFormatterInterface $formater
-     */
-    public static function setDefaultDataFormatter(DataFormatterInterface $formater)
-    {
-        self::$defaultDataFormatter = $formater;
-    }
-
-    /**
-     * Returns the default data formater
-     *
-     * @return DataFormatterInterface
-     */
-    public static function getDefaultDataFormatter()
-    {
-        if (self::$defaultDataFormatter === null) {
-            self::$defaultDataFormatter = new DataFormatter();
-        }
-        return self::$defaultDataFormatter;
-    }
-
-    /**
-     * Sets the data formater instance used by this collector
-     *
-     * @param DataFormatterInterface $formater
-     * @return $this
-     */
-    public function setDataFormatter(DataFormatterInterface $formater)
-    {
-        $this->dataFormater = $formater;
-        return $this;
-    }
-
-    /**
-     * @return DataFormatterInterface
-     */
-    public function getDataFormatter()
-    {
-        if ($this->dataFormater === null) {
-            $this->dataFormater = self::getDefaultDataFormatter();
-        }
-        return $this->dataFormater;
-    }
 
     /**
      * Shorten the file path by removing the xdebug path replacements
@@ -132,79 +84,6 @@ abstract class DataCollector implements DataCollectorInterface
         if ($url) {
             return ['url' => $url, 'ajax' => $this->getXdebugShouldUseAjax()];
         }
-    }
-  
-    /**  
-     * Sets the default variable dumper used by all collectors subclassing this class
-     *
-     * @param DebugBarVarDumper $varDumper
-     */
-    public static function setDefaultVarDumper(DebugBarVarDumper $varDumper)
-    {
-        self::$defaultVarDumper = $varDumper;
-    }
-
-    /**
-     * Returns the default variable dumper
-     *
-     * @return DebugBarVarDumper
-     */
-    public static function getDefaultVarDumper()
-    {
-        if (self::$defaultVarDumper === null) {
-            self::$defaultVarDumper = new DebugBarVarDumper();
-        }
-        return self::$defaultVarDumper;
-    }
-
-    /**
-     * Sets the variable dumper instance used by this collector
-     *
-     * @param DebugBarVarDumper $varDumper
-     * @return $this
-     */
-    public function setVarDumper(DebugBarVarDumper $varDumper)
-    {
-        $this->varDumper = $varDumper;
-        return $this;
-    }
-
-    /**
-     * Gets the variable dumper instance used by this collector; note that collectors using this
-     * instance need to be sure to return the static assets provided by the variable dumper.
-     *
-     * @return DebugBarVarDumper
-     */
-    public function getVarDumper()
-    {
-        if ($this->varDumper === null) {
-            $this->varDumper = self::getDefaultVarDumper();
-        }
-        return $this->varDumper;
-    }
-
-    /**
-     * @deprecated
-     */
-    public function formatVar($var)
-    {
-        return $this->getDataFormatter()->formatVar($var);
-    }
-
-    /**
-     * @deprecated
-     */
-    public function formatDuration($seconds)
-    {
-        return $this->getDataFormatter()->formatDuration($seconds);
-    }
-
-    /**
-     * @deprecated
-     */
-    public function formatBytes($size, $precision = 2)
-    {
-        return $this->getDataFormatter()->formatBytes($size, $precision);
     }
 
     /**
